Bug 2087232: last namespace instead of active namespace
